# Add arm64 to GOARCH switches